requirements.txt
solders==0.21.0
python-dotenv==1.0.1
solana==0.34.3
requests